Width/Height of levels made not constant
Updated LevelMap so that the height and width of a level aren't constant. Added getter functions for height and width. Also updated the file format to contain the width and height before the tile data. The constructor using an array will not be updated as if a variable sized level is needed, then a file should be used