Player lives
Added a lives system, where on death, the player loses a life and becomes invulnerable for a few seconds. During this time, the opacity of the player varies to show that they are in this state. If the player lives drops to 0, then they are dead, so this part works the same as before