Death sound effects
Created SoundList singleton class which contains a list of soundeffect pointers. When a CharacterEnemy dies, if it has a death sound it adds it to the SoundList during deletion. This way, the sound can play out in full after the enemy has died.
Also added sounds for coins and koopas