Level 2 update
Added Level2 assets.
Added a bool to turn off screenwrapping
Added an xOffset parameter in CharacterPlayable's render
Updated how passable tiles work, so that it is a variable of the current tileset.
Updated Level2, so that it loads the level and renders it. It also creates a vector of impassable tiles, and creates a vector of passable tiles from that