Level 2 death/lives update
Added a black version of the font so that it's easier to read on level 2.
Updated CharacterEnemy so that the death sound is only played if the enemy is actually dead, not just being deleted.
Updated CharacterMario so that if the player is below the level, they die and so that when they die, the level is reset.
Updated GameScreenLevel2 so that setting up the level and enemies has been extracted to functions. Additionally, a RestartLevel function has been added which gets called when the player dies